+++ CHANGELOG.md
@@ -34,7 +34,11 @@ creating or updating the phone number.
 - Order of events and acknowledgements in feed in detailed incident view. Order is now oldest-first.
 - Made docker-compose build more flexible.
 
+- Make it easier/more obvious to save a profile without a phone number by selecting the option "None".
+- Option "None" is default in the phone number selector in notification profiles.
+
 - Updated the phonenumber component to have prettier and more sensible css.
+
 
 
 ### Added
@@ -96,5 +100,6 @@ creating or updating the phone number.
 - Success/error alerts on all bulk-operations on incidents (ack, add-ticket-url, open, close).
 - The ability to filter out older incidents via Timeframe selector.
 - Display of frontend-, backend- and API versions below the incidents table.
+